QSystemTrayIcon: initialize all members
The 'showArrow' member was not init'ed.

Initialize it to true, which is the default value of the
QBalloonTip::showBalloon() function's argument of the same
purpose.

Reported as new by Coverity, but dating back all the way to
cc3875c2e463be5cf126a18637295a0c56358eda, so affects all
current branches.

Coverity-Id: 171482
Change-Id: Ica519ecda3a4ae413f606faab8c22f7072f412a8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>