QTzTimeZonePrivate: resolve a misguided TODO
A three-way if/else-if/else had the same line of code as all three
branches and a TODO to fix that.  As it happens, the flags being
tested here are irrelevant; they indicate whether the transition time
(which is always given in UTC) was *specified* (i.e. the how a
relevant authority identified the transition time) in terms of local
wall-clock time, local standard time or UTC.  The correction
contemplated by TODO (and experimented with in an earlier version of
this change, that broke correct tests) has in fact been done for us by
the zoneinfo package's zic (zone-info compiler) in the course of
writing the binary file we're parsing.  These flags are only present
in the binary file to enable the date command to correctly handle
POSIX-style values for the TZ environment variable.

We consequently have no need for the tz_ttisgmt or tz_ttisstd fields
of our QTzType and can save the bother of recording them, when reading
their part of the file.

Change-Id: Ia33e87291ecc383eb5cb796d7b8a5213a94f1648
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>